requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
setuptools==70.0.0
simplejwt==2.0.1
sqlparse==0.5.0
typing==3.7.4.3
uritemplate==4.1.1
